examples/alibaba-cloud/requirements.txt
pipecat-ai[silero,websocket]
python-dotenv
openai
websockets
